added authorizer for edit/delete permission

+++ Back/routes/travels/deleteTravel.js
@@ -1,8 +1,14 @@
+const authorizer = require("../../misc/authorizer");
 const errors = require("../../misc/errors");
 const prisma = require("../../prisma");
 
 module.exports = () => async (req, res, next) => {
     const { id } = req.params;
+    const user = res.locals;
+
+    const authorized = await authorizer(id, user);
+
+    if(!authorized) return next(errors[401]);
 
     await prisma.services.deleteMany({
         where: {
